AddRating logic finished, with cases not rated, not rated by user and already rated. getUsers API Call to be implemented
