+++ reactApp/src/data/getUsers.js
@@ -1,0 +1,13 @@
+
+function getUsers() {
+    // HARD-CODED FOR TESTINGS AND DEMO PURPOSES. 
+    // To be replaced by the API call to get the user ID
+    const UserId = "6420450b3d2951c719ec768"
+
+  return (
+    UserId
+    )
+}
+
+export default getUsers
+
